remove DateChip file
