--- app/src/main/java/com/sanaa/tudee_assistant/presentation/designSystem/component/DateChip.kt
@@ -1,43 +1,0 @@
-package com.sanaa.tudee_assistant.presentation.designSystem.component
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Icon
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.unit.dp
-import com.sanaa.tudee_assistant.R
-import com.sanaa.tudee_assistant.presentation.designSystem.theme.Theme
-
-@Composable
-fun DateChip(date: String) {
-    Row(
-        modifier = Modifier
-            .clip(RoundedCornerShape(100.dp))
-            .background(Theme.color.surface)
-            .padding(vertical = 6.dp, horizontal = Theme.dimension.small),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(2.dp)
-    ) {
-        Icon(
-            modifier = Modifier.size(Theme.dimension.regular),
-            painter = painterResource(id = R.drawable.calendar_favorite_01),
-            contentDescription = null,
-            tint = Theme.color.body
-        )
-
-        Text(
-            text = date,
-            color = Theme.color.body,
-            style = Theme.textStyle.label.small
-        )
-    }
-}
