+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/designSystem/component/TaskItemCard.kt
@@ -27,6 +27,7 @@ import com.sanaa.tudee_assistant.presentation.composable.CategoryThumbnail
 import com.sanaa.tudee_assistant.presentation.designSystem.theme.Theme
 import com.sanaa.tudee_assistant.presentation.designSystem.theme.TudeeTheme
 import com.sanaa.tudee_assistant.presentation.model.TaskUiState
+import com.sanaa.tudee_assistant.presentation.screen.categoryTask.components.DateChip
 import com.sanaa.tudee_assistant.presentation.utils.DataProvider
 
 @Composable
